% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{class}
\name{cellData-class}
\alias{[,cellData,character,ANY,ANY-method}
\alias{cellData}
\alias{cellData,cellData-class}
\alias{cellData,cellData-method}
\alias{cellData-class}
\alias{show,cellData-method}
\title{The cellData class}
\usage{
cellData(name, ctrwell, expwell = NULL, ...)

\S4method{[}{cellData,character,ANY,ANY}(x, i)

\S4method{show}{cellData}(object)
}
\description{
The main class used in \code{OperaMate} to hold all levels of experiment
data of a specific type.
}
\section{Slots}{

\describe{
\item{\code{name}}{character, one parameter in the Columbus system report.}

\item{\code{ctrwell}}{a character vector, the control well IDs, e.g. B05.}

\item{\code{expwell}}{a character vector, the sample well IDs, e.g. C15.}

\item{\code{origin.data}}{a numeric matrix,
the raw data matrix with rows the well IDs and columns the plate IDs.}

\item{\code{norm.data}}{a numeric matrix, the normalized data.}

\item{\code{qc.data}}{a numeric matrix, the data after quality
control, with the rows are "barcode:wellID" and columns are the data of
all replicated samples and their means, the p-values in quality control
and if they have passed the quality control.}

\item{\code{norm.method}}{character the normalization method.}

\item{\code{qc.th}}{numeric, the threshold in the quality control.}

\item{\code{plate.quality}}{a numeric matrix, the quality data with
the rows are the barcode and columns are the replicateIDs.}

\item{\code{Sig}}{a list of the following components:
\itemize{
\item{\code{SigMat}:}{a logic matrix marking the high and low expressed hits}
\item{\code{threshold}:}{the threshold of the high and low expressed hits}
\item{\code{stats}:}{the numbers of the high and low expressed hits}
\item{\code{pvalue}:}{the pvalue of each sample by t tests}
}}
}}
\section{Methods}{

\describe{
\item{Constructor}{
\code{cellData(name, ctrwell, expwell = NULL)}. \code{name}: character,
one parameters in the Columbus system reports.
}
\item{Show}{signature(object="cellLoad"). Displays object content as text.}
\item{Accessor}{x[i]. \code{x}: a cellData object;
\code{i}: character, a cellData slot name.}
}
}
\examples{
oneCell <- cellData(name="Average Intensity of Nuclei",
ctrwell = c("C23","E23","G23"))
oneCell
oneCell["name"]
}

