% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{methods}
\name{cellLoad}
\alias{cellLoad}
\alias{cellLoad,cellData-method}
\title{Data importing}
\usage{
cellLoad(object, lstPlates)

\S4method{cellLoad}{cellData}(object, lstPlates)
}
\arguments{
\item{object}{a cellData object}

\item{lstPlates}{a list of expData objects}
}
\value{
a \code{cellData} object, with initialized slot \code{origin.data}
}
\description{
Extracts data of a specific type in a list of \code{expData} objects to
initialize a \code{cellData} object.
}
\examples{
data(platemap)
platemap$path <- file.path(system.file("Test",package = "OperaMate"),
platemap$path)
lstPlates <- loadAll(cellformat="Tab",datapath=dirname(platemap$path[1]))
data(demoCell)
oneCell <- cellLoad(oneCell, lstPlates)
str(oneCell["origin.data"])
}

