% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{methods}
\name{cellMean}
\alias{cellMean}
\alias{cellMean,cellData,cellData,character-method}
\title{Mean of two cellData objects}
\usage{
cellMean(cell1, cell2, name)

\S4method{cellMean}{cellData,cellData,character}(cell1, cell2, name)
}
\arguments{
\item{cell1}{one cellData object}

\item{cell2}{another celldata object}

\item{name}{the name of mean cellData object}
}
\value{
the mean cellData object
}
\description{
Merges the intensities in nucleus and cytoplasm to their averages for
signature detection.
}
\examples{
data(demoCell)
meanCell <- cellMean(oneCell, oneCell, "meanCell")
meanCell
}

