% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{methods}
\name{cellNorm}
\alias{cellNorm}
\alias{cellNorm,cellData-method}
\title{Data normalization}
\usage{
cellNorm(object, norm.method)

\S4method{cellNorm}{cellData}(object,
  norm.method = getOption("opm.normalization.method"))
}
\arguments{
\item{object}{a cellData object}

\item{norm.method}{getOption("opm.normalization.method")}
}
\value{
a \code{celldata} object with initialized slot \code{norm.data}
}
\description{
Normalizes raw data based on different normalization methods.
}
\details{
Method description: "MP" employes the median polish algorithm which
divides data by the median of their plates and wells recursively, while
"PMed" only divides data by the median of their plates; "Z" substracts
data by their plate medians, and then divides by the median absolute
deviations; "Ctr" divides data by the mean of their plate negative controls;
"None" avoids the data normalization in this step. The first three
methods are based on the assumption that most samples display no
biological effects in the assay be analyzed. They are often more
effective than "Ctr" method as to the high throughput screening.
}
\examples{
data(demoCell)
oneCell <- cellNorm(oneCell, norm.method = "MP")
str(oneCell["norm.data"])
}

