% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{methods}
\name{cellQC}
\alias{cellQC}
\alias{cellQC,cellData-method}
\title{Quality control}
\usage{
cellQC(object, qth = 0.05, outpath = "./", gDevice = "dev.new", ...)

\S4method{cellQC}{cellData}(object, qth = 0.05, outpath = "./",
  gDevice = "dev.new", ...)
}
\arguments{
\item{object}{a cellData object}

\item{qth}{a numeric threshold of pvalues.}

\item{outpath}{a character string naming the location the figure to
generate to.}

\item{gDevice}{the graphic device (default: dev.new).
gDevice=NA generates no figure.}

\item{...}{arguments of the graphic device}
}
\value{
a \code{cellData} object with intialized slot \code{qc.data}
and \code{plate.quality}
}
\description{
Checks quality of all plates and then wells.
}
\details{
Requires three or more replicated samples.
}
\examples{
data(demoCell)
oneCell <- cellQC( oneCell, qth = .05, gDevice = NA)
str(oneCell["qc.data"])
str(oneCell["plate.quality"])
}

