% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{methods}
\name{cellSig}
\alias{cellSig}
\alias{cellSig,cellData-method}
\title{Hit identification}
\usage{
cellSig(object, method = c("stable", "ksd", "kmsd"), th = NULL,
  thPval = 0.05, digits = 3, adjust.method = p.adjust.methods,
  plot = TRUE, outpath = getOption("opm.outpath"), ...)

\S4method{cellSig}{cellData}(object, method = c("stable", "ksd", "kmsd"),
  th = NULL, thPval = 0.05, digits = 3,
  adjust.method = p.adjust.methods, plot = TRUE,
  outpath = getOption("opm.outpath"), ...)
}
\arguments{
\item{object}{a cellData object}

\item{method}{method = c("stable","ksd","kmsd").
Details are referred in the vignette.}

\item{th}{numeric, the thresholds. It can be one threshold  for both
high and low expressed hit or two thresholds for each respectively.}

\item{thPval}{numeric, threshold of pvalues in the t-test
between the sample and control replicates}

\item{digits}{integer, the number of digits used to show the thresholds}

\item{adjust.method}{pvalue correction method}

\item{plot}{plot QQ-plot when method is "stable" if TRUE.}

\item{outpath}{directory of output figures, default: getOption("opm.outpath")}

\item{...}{arguments of the graphic device}
}
\value{
a \code{cellData} object with initialized slot \code{Sig}.
}
\description{
Detects samples those are most different from the negative controls.
}
\examples{
data(demoCell)
op <- options("device")
options("device" = "png")
oneCell <- cellSig(oneCell, method = "stable", th = c(0.05, 0.05),
outpath = tempdir())
options(op)
names(oneCell["Sig"])
}

