% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-methods.R
\docType{methods}
\name{cellSigAnalysis}
\alias{cellSigAnalysis}
\title{Hits function analysis}
\usage{
cellSigAnalysis(object, genemap, organism, type = c("High", "Low"),
  file = NULL, ...)
}
\arguments{
\item{object}{a cellData object}

\item{genemap}{a data frame, the well-gene specification table}

\item{organism}{organism name.}

\item{type}{include both high and low expressed hits or one of them.}

\item{file}{the filename of the enrichment table (default: disabled)}

\item{...}{the arguments of \code{gprofiler}.}
}
\value{
a data frame of the functional report from gProfiler
}
\description{
Performs function analysis using gProfileR
}
\details{
genemap must include colnames "Barcode","Well","GeneSymbol".
organism name can be referred to g:Profiler tool.
For example, human: hsapiens, mouse: mmusculus.
}
\examples{
data(demoCell)
genemap <- read.csv(file.path(system.file("demoData", package = "OperaMate"),
"genemap.csv"), stringsAsFactors = FALSE)
chart <- cellSigAnalysis(oneCell, genemap, organism = "mmusculus")
head(chart)
}

