% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{cellSigAnalysisPlot}
\alias{cellSigAnalysisPlot}
\title{The barplot of enrichment functions}
\usage{
cellSigAnalysisPlot(chart, prefix = "", type = NULL, fill = "steelblue",
  outpath = getOption("opm.outpath"), ...)
}
\arguments{
\item{chart}{data frame, the functional annotation chart}

\item{prefix}{character, the prefix of figure name}

\item{type}{selected domains from chart, e.g. BP.}

\item{fill}{color of the bars}

\item{outpath}{directory of output figures, default: getOption("opm.outpath")}

\item{...}{other arguments for graphical devices}
}
\value{
Invisibly the ggplot2 function for barplot
}
\description{
The barplot of enrichment functions
}
\examples{
data(demoCell)
genemap <- read.csv(file.path(system.file("Test", package = "OperaMate"),
"demoData", "genemap.csv"), stringsAsFactors = FALSE)
chart <- cellSigAnalysis(oneCell, genemap, organism = "mmusculus")
op <- options("device")
options("device" = "png")
cellSigAnalysisPlot(chart, type = "BP", outpath = tempdir())
options(op)
}

