% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-methods.R
\docType{methods}
\name{cellSigplot}
\alias{cellSigplot}
\title{Hits volcano plot}
\usage{
cellSigplot(object, outpath = "./", gDevice = "dev.new",
  color.highlight = "red", color.background = "blue",
  highlight.label = NULL, ...)
}
\arguments{
\item{object}{a cellData object}

\item{outpath}{a character string naming the location the figure to
generate to}

\item{gDevice}{the graphic device (default: dev.new)}

\item{color.highlight}{a character specifying the color of the hits}

\item{color.background}{a character specifying the color of the other samples}

\item{highlight.label}{a vector of characters specifying the names of the
samples to be highlighted, with the names are the "barcode:wellID".}

\item{...}{arguments of the graphic device}
}
\value{
Invisibly an object of \code{ggplot}
}
\description{
Visualizes hits by volcano plot.
}
\details{
Users can highlight a certain samples during plottting.
}
\examples{
data(demoCell)
labels <- c("Axin1")
names(labels) <- c("DSIMGA04:C07")
cellSigplot(oneCell,highligh.label=labels)
}

