% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-methods.R
\docType{methods}
\name{cellViz}
\alias{cellViz}
\title{Batch effect visualization}
\usage{
cellViz(object, exps = c("all", "exp", "ctr"), type = c("heatmap", "median",
  "PCA"), gDevice = "dev.new", outpath = "./", ifCompare = TRUE,
  control.par = list(), ...)
}
\arguments{
\item{object}{a cellData object}

\item{exps}{exps = c("all", "exp","ctr"), "exp" and "ctr" consider only the
sample and control wells respectively.}

\item{type}{type = c("heatmap","median","PCA")}

\item{gDevice}{the graphic device (default: dev.new)}

\item{outpath}{a character string naming the location the figures to generate
to}

\item{ifCompare}{logic, comparison before and after normalization will be
shown if the data normalization has been processed.}

\item{control.par}{list, the parameter to be passed to plotting methods}

\item{...}{arguments of the graphic device}
}
\value{
Invisibly a list of two components:
\describe{
\item{\code{g1}:}{Plotting results before data normalization.}
\item{\code{g2}:}{Plotting results after data normalization.
g2=NULL if ifCompare=FALSE.}
}
}
\description{
Visualizes batch effects by three ways: hieratical clustering (heatmap),
the principal component analysis (PCA) and the median scatter plot (median).
}
\details{
Heatmap: a large region of distinguishing color indicates the systematical
biases; PCA: points distant from others indicate data with excessive noises;
Median: the distribution of the plate and well medians.
}
\section{control.par}{

type="heatmap"
\describe{
\item{\code{    fontsize}:}{the fontsize of the plot (default: 15)}
\item{\code{    color}:}{vector of character specifying the colors used in
heatmap}
\item{\code{    fontsize_row}:}{fontsize of rownames (default: 0)}
\item{\code{    fontsize_col}:}{fontsize of colnames (default: 0)}
\item{\code{    xlab}:}{character, X axis label}
\item{\code{    ylab}:}{character, Y axis label}
\item{\code{    remove.noise}:}{logical. If TRUE, the heatmap only represents
data in the range of [median-3mad, median+3mad]}
\item{\code{    ...}:}{arguments to be passed to \code{pheatmap}}
}

type="median"
\describe{
\item{\code{    lwd}:}{the line widths}
\item{\code{    pch}:}{the plotting symbols}
\item{\code{    main}:}{the title of the figure}
\item{\code{    ...}:}{other parameters for function \code{plot}}
}

type="PCA"
\describe{
\item{\code{    main}:}{the title of the figure}
\item{\code{    pch}:}{the plotting symbols of the legends}
\item{\code{    ...}:}{other parameters for function \code{plot}}
}
}
\examples{

data(demoCell)
cellViz(oneCell,exps="exp",type="heatmap",ifCompare=FALSE)
cellViz(oneCell,exps="exp",type="median",ifCompare=FALSE)
cellViz(oneCell,exps="exp",type="PCA")
}

