% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{demoData}
\alias{demoData}
\alias{onCellNum}
\alias{oneCell}
\alias{oneCellNum}
\alias{platemap}
\title{Examples of tables and cellData objects}
\value{
platemap: a data frame

oneCell: a \code{cellData} object

oneCellNum: a \code{cellData} object
}
\description{
oneCellNum
}
\section{platemap}{

\describe{
\item{Description}{The experiment information of each Columbus
analysis report. This table is required only if the report formats are not
standarded. See \code{\link{loadAll}} for more information.}
\item{Format}{data.frame with the following required column names:
\describe{
\item{\code{FileName}:}{character, the name of the report.}
\item{\code{Format}:}{character, only ''Tab'' and ''Matrix'' are supported in
the current version.}
\item{\code{Barcode}:}{character, the barcode of the plates.}
\item{\code{RepID}:}{character, the ID to distinguish the replicated plates.}
\item{\code{Path}:}{character, the full path of the report.}
}
}
}
}

\section{oneCell}{

\describe{
\item{Description}{
\code{oneCell} is a \code{cellData} object used in the examples of the
package.}
}
}

\section{oneCellNum}{

\describe{
\item{Description}{
\code{oneCellNum} is a \code{cellData} object storing the cell numbers.}
}
}
\examples{
data(platemap)
str(platemap)
data(demoCell)
oneCell
data(demoCellNum)
}
\keyword{data}

