% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData-class.R
\docType{class}
\name{expData-class}
\alias{[,}
\alias{[,expData,ANY,ANY-method}
\alias{dataLoad}
\alias{dataLoad,}
\alias{dataLoad,expData-method}
\alias{expData}
\alias{expData,}
\alias{expData-class}
\alias{expData-method}
\alias{show,expData-method}
\title{The expData class}
\usage{
expData(name, path, rep.id, exp.id, format)

\S4method{show}{expData}(object)

\S4method{[}{expData,ANY,ANY}(x, i)

dataLoad(object, data, wellID)

\S4method{dataLoad}{expData}(object, data, wellID)
}
\arguments{
\item{name}{character, the plate ID (barcode-replicateID), e.g. DSIMGA03-s1.}

\item{path}{character, the path of the Columbus system report.}

\item{rep.id}{character, replicateID, e.g. s1.}

\item{exp.id}{character, barcode, e.g. DSIMGA03.}

\item{format}{character, format of the Columbus system report.}

\item{object}{a expData class}

\item{x}{a expData object}

\item{i}{a requested slot name}

\item{data}{the vectorized raw data matrix of one plate of each type.}

\item{wellID}{a character vector, the well IDs.}
}
\value{
an expData object
}
\description{
The expData class is a container to store data imported from one
Columbus system report

Constructor method of expData class.

Show method
}
\section{Slots}{

\describe{
\item{\code{name}}{character, the plate ID (barcode-replicateID), e.g. DSIMGA03-s1.}

\item{\code{path}}{character, the path of the Columbus system report.}

\item{\code{rep.id}}{character, replicateID, e.g. s1.}

\item{\code{exp.id}}{character, barcode, e.g. DSIMGA03.}

\item{\code{data}}{a list of vectors,
the vectorized raw data matrix of one plate of each type.}

\item{\code{format}}{character, format of the Columbus system report.}

\item{\code{wellID}}{a character vector, the well IDs.}
}}
\section{Methods}{
 
\describe{
\item{Constructor}{
\code{expData(name, path, rep.id, exp.id, format)}.
}
\item{Show}{signature(object = "expData"). Displays object content as text.}
\item{Accessor}{x[i]. \code{x}: an expData object;
\code{i}: character, an expData slot name.}
\item{dataLoad}{\code{dataLoad(object, data, wellID)}}
}
}
\examples{
onePlate <- expData(name = "130504-s1-02.txt",
         path = file.path(system.file("Test", package = "OperaMate"),
                           "Matrix", "130504-s1-02.txt"),
       rep.id = "s1",
       exp.id = "DSIMGA02",
       format = "Matrix")
onePlate
onePlate["name"]
}

