% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GenerateReport}
\alias{GenerateReport}
\title{Report generation}
\usage{
GenerateReport(lstCells, genemap, file = NULL, outpath = "./",
  gDevice = "dev.new", ...)
}
\arguments{
\item{lstCells}{a list of cellData objects}

\item{genemap}{the mapping file of all samples}

\item{file}{the path of the file to generate to}

\item{outpath}{a character string naming the location the figures to
generate to}

\item{gDevice}{the graphic device (default: dev.new).
gDevice=NA generates no figure}

\item{...}{arguments of the graphic device}
}
\value{
a data frame with annotated information of each well
}
\description{
Summarizes all results in the list of \code{cellData} objects,
and writes out a report to file.
}
\details{
This function summarizes the information from all \code{cellData} objects,
and visualizes the number of the hists if required.
}
\examples{
data(demoCell)
data(genemap)
anno.data <- GenerateReport(list(oneCell),genemap,gDevice=NA)
colnames(anno.data)

}

