% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nameParser}
\alias{nameParser}
\title{Plate information extraction}
\usage{
nameParser(vec.files, egFilename)
}
\arguments{
\item{vec.files}{a vector of file names}

\item{egFilename}{a file name example}
}
\value{
a data frame of PlateID, RepID, and Barcode
}
\description{
Extract plate information from file names.
}
\details{
An example of egFilename = list(eg.filename = "0205-s2-01.txt",
rep.id = "s2", exp.id = "01", sep = "-", barcode = "DSIMGA01").
}

