% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{operaMate}
\alias{operaMate}
\title{Data process and analysis pipeline}
\usage{
operaMate(cellformat = c("Tab", "Matrix", "Others"), datapath = "./",
  platemap = NULL, genemap = NULL,
  cellnames = c("Average.Intensity.of.Nuclei",
  "Average.Intensity.of.Cytoplasm", "Average.Total.Intensity",
  "Average.Intensity.Ratio"), ctrwell = NULL, expwell = NULL,
  Prefix.Barcode = "DSIMGA", norm.method = c("Both", "Plate", "Z", "Ctr",
  "None"), gDevice = "png", outpath = "./", BatcheffectSample = c("all",
  "exp", "ctr"), BatcheffectType = c("heatmap", "median", "PCA"),
  BatcheffectifCompare = TRUE, viz.par = list(), qth = 0.05,
  Sig.method = c("ktsd", "ksd", "kmsd"), Sig.threshold = NULL,
  threshold.pvalue = 0.05, anno.file = file.path(outpath,
  "OperaMateReport.txt"), email = NA, david.type = c("all", "high", "low"),
  david.terms = NULL, david.count = 2L, david.threshold = 0.1,
  david.termN = NA, ...)
}
\arguments{
\item{cellformat}{character specifying the format of the reports}

\item{datapath}{character specifying the location of the reports}

\item{platemap}{data frame, required when cellformat = "Others".
See an example as \code{\link{platemap}}.}

\item{genemap}{data frame, the mapping file of all samples.
See an example as \code{\link{genemap}}.}

\item{cellnames}{vector of character specifying the data types.
The types are the parameters in the Columbus system reports.}

\item{ctrwell}{a vector of characters specifying the control well IDs}

\item{expwell}{a vector of characters specifying the control sample IDs}

\item{Prefix.Barcode}{character, the Barcode prefix}

\item{norm.method}{norm.method = c("Both","Plate","Z","Ctr","None")}

\item{gDevice}{the graphic device (default: png)}

\item{outpath}{a character string naming the location the figures to
generate to}

\item{BatcheffectSample}{c("all", "exp", "ctr"), "exp" and "ctr" consider
only the sample and control wells respectively.}

\item{BatcheffectType}{c("heatmap","median","PCA-plate","PCA-well")}

\item{BatcheffectifCompare}{logic, comparison before and after normalization
will be shown if the data normalization has been processed.}

\item{viz.par}{list, arguments to be passed to \code{\link{cellViz}}}

\item{qth}{numeric, the threshold in quality control}

\item{Sig.method}{c("ktsd","ksd","kmsd")}

\item{Sig.threshold}{numeric, the threshold for hit identification}

\item{threshold.pvalue}{numeric, threshold of pvalues in the t-test between
the sample and control replicates}

\item{anno.file}{the path to generate the report}

\item{email}{the email for the DAVID webservice registration}

\item{david.type}{c("all","high","low")}

\item{david.terms}{a list of DAVID annotation categories}

\item{david.count}{numeric, number of items}

\item{david.threshold}{numeric, threshold in the DAVID analysis}

\item{david.termN}{numeric, number of terms to show in the figure}

\item{...}{arguments of the graphic device}
}
\value{
a list of two components:  a list of cellData objects and  the
annotated table of each well
}
\description{
A systematical pipeline for opera data importing, normalization, quality
control, signature gene detection and visualization.
}
\examples{
data(platemap); data(genemap)
platemap$path <- file.path(system.file("Test",package = "OperaMate"),
                           platemap$path)
(outpath <- tempdir())
operaReport <- operaMate(cellformat = "Others",
platemap = platemap,genemap = genemap,
Prefix.Barcode = "DSIMGA",outpath=outpath,
expwell = paste0(rep(LETTERS[2:15],each=20),
rep(formatC(3:22,width=2,flag=0),times=14)),
ctrwell = c("C23","E23","G23"),
email = "cliu@sjtu.edu.cn",
norm.method="Both",BatcheffectSample="exp",
Sig.method="ktsd",david.type="all",
width=960,height=960)
head(operaReport$data.anno)

}

