% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData-methods.R
\docType{methods}
\name{parseTemplete}
\alias{parseTemplete}
\title{Data extraction from one report}
\usage{
parseTemplete(onePlate, well.digits = 2)
}
\arguments{
\item{onePlate}{an expData object}

\item{well.digits}{the digits of the well column in the well-gene
specification file}
}
\value{
an \code{expData} object with initialized slot \code{data}.
}
\description{
Extracts data in the report to the slot \code{data} in the \code{expData}
object. An inner function of \code{\link{loadAll}}.
}
\examples{
datapath <- file.path(system.file("Test", package = "OperaMate"), "Tab")
lstPlates <- loadAll(cellformat = "Tab", datapath = datapath )
onePlate <- parseTemplete(lstPlates[[1]])
}

