% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cellData-methods.R
\docType{methods}
\name{plateViz}
\alias{plateViz}
\title{One plate visualization}
\usage{
plateViz(object, ID, gDevice = "dev.new", outpath = "./", ...)
}
\arguments{
\item{object}{a cellData object}

\item{ID}{character, the plate ID to be visualized}

\item{gDevice}{the graphic device (default: dev.new),
gDevice=NA generates no figure}

\item{outpath}{a character string naming the location the figure to
generate to}

\item{...}{arguments to be passed to \code{pheatmap}}
}
\value{
Invisibly a list of two components:
\describe{
\item{\code{g1}:}{a list of component by \code{pheatmap} based on data
before normalization}
\item{\code{g2}:}{a list of component by \code{pheatmap} based on data
after normalization}
}
}
\description{
Visualizes data of one plate using heatmap.
}
\examples{
data(demoCell)
plateViz(oneCell,ID="DSIMGA04-s2")
}

