% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{methods}
\name{cellLoad}
\alias{cellLoad}
\alias{cellLoad,cellData-method}
\title{Data importing}
\usage{
cellLoad(object, lstPlates, ...)

\S4method{cellLoad}{cellData}(object, lstPlates, positive.ctr = NULL,
  negative.ctr = NULL, neglect.well = NULL, expwell = NULL)
}
\arguments{
\item{object}{a cellData object}

\item{lstPlates}{a list of expData objects}

\item{...}{other parameters}

\item{positive.ctr}{a character vector, the positive control well IDs,
e.g. c("E05", "E06")}

\item{negative.ctr}{a character vector, the positive control well IDs,
e.g. c("B05", "B06")}

\item{neglect.well}{a character vector, the neglect wells.
Accept regular expression, e.g. c("*02", "*23")}

\item{expwell}{include all wells except control and neglect.well if NULL}
}
\value{
a \code{cellData} object, with initialized slot \code{origin.data}
}
\description{
Extracts data of a specific type in a list of \code{expData} objects to
initialize a \code{cellData} object.
}
\details{
negative.ctr accept regular expression
}
\examples{
data(platemap)
platemap$Path <- file.path(
system.file("Test", package = "OperaMate"), platemap$Path)
data(demoCell)
datapath <- file.path(system.file("Test", package = "OperaMate"), "Matrix")
lstPlates <- loadAll(cellformat = "Matrix", datapath = datapath)
oneCell <- cellLoad(oneCell, lstPlates, neglect.well = c("*02", "*23"))
str(oneCell["origin.data"])
}

