% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{methods}
\name{cellQC}
\alias{cellQC}
\alias{cellQC,cellData-method}
\title{Quality control}
\usage{
cellQC(object, qcType = NULL, qc.threshold = NULL,
  replace.badPlateData = TRUE, plot = TRUE,
  outpath = getOption("opm.outpath"), ...)

\S4method{cellQC}{cellData}(object, qcType = getOption("opm.QC.type"),
  qc.threshold = getOption("opm.QC.threshold"),
  replace.badPlateData = getOption("opm.replace.badPlateData"), plot = TRUE,
  outpath = getOption("opm.outpath"), ...)
}
\arguments{
\item{object}{a cellData object}

\item{qcType}{the type of quality control}

\item{qc.threshold}{quality control thresholds}

\item{replace.badPlateData}{if TRUE,
replace the values of bad plate by their replicates}

\item{plot}{if TRUE, plot figures}

\item{outpath}{directory of output figures, default: getOption("opm.outpath")}

\item{...}{arguments for the graphic device}
}
\value{
a \code{cellData} object with intialized slot \code{qc.data},
\code{plate.quality} and \code{plate.quality.data}.
}
\description{
Checks quality of all plates and then wells.
}
\details{
Requires three or more replicated samples.

qcType include c("plateCorrelation", "wellSd", "zFactor", "cellNumber"),
An example of qc.threshold is c(correlation = 0.8, zfactor = 0.5, cellnumber = 50).
}
\examples{
data(demoCell)
op <- options("device")
options("device" = "png")
oneCell <- cellQC(oneCell, qcType = c("plateCorrelation", "wellSd", "cellNumber"),
qc.threshold = c(correlation = 0.7), outpath = tempdir())
options(op)
str(oneCell["qc.data"])
str(oneCell["plate.quality"])
}

