% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-class.R
\docType{methods}
\name{cellSig}
\alias{cellSig}
\alias{cellSig,cellData-method}
\title{Hit identification}
\usage{
cellSig(object, method = c("ktsd", "ksd", "kmsd"), threshold.method = NULL,
  digits = 3, outpath = "./", gDevice = "dev.new",
  threshold.pvalue = 0.05, adjust.method = p.adjust.methods, ...)

\S4method{cellSig}{cellData}(object, method = c("ktsd", "ksd", "kmsd"),
  threshold.method = NULL, digits = 3, outpath = "./",
  gDevice = "dev.new", threshold.pvalue = 0.05,
  adjust.method = p.adjust.methods, ...)
}
\arguments{
\item{object}{a cellData object}

\item{method}{method=c("ktsd","ksd","kmsd").
Details are referred in the vignette.}

\item{threshold.method}{numeric}

\item{digits}{integer, the number of digits used to show the results}

\item{outpath}{a character string naming the location the figure to
generate to. Only for method = "ktsd".}

\item{gDevice}{the graphic device (default: dev.new),
gDevice=NA generates no figure.  Only for method = "ktsd".}

\item{threshold.pvalue}{numeric, threshold of pvalues in the t-test
between the sample and control replicates}

\item{adjust.method}{pvalue correction method}

\item{...}{arguments of the graphic device}
}
\value{
a \code{cellData} object with initialized slot \code{Sig}.
}
\description{
Detects samples those are most different from the negative controls.
}
\examples{
data(demoCell)
oneCell <- cellSig( oneCell, method = "ktsd",gDevice=NA )
names(oneCell["Sig"])

}

