% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-methods.R
\docType{methods}
\name{cellSigAnalysis}
\alias{cellSigAnalysis}
\title{Hits function analysis}
\usage{
cellSigAnalysis(object, genemap, email, type = c("all", "high", "low"),
  david.terms = NULL, count = 2L, threshold = 0.1, showTermNum = NA,
  showTermBy = "PValue", showTermCol = "blue", file = NA,
  outpath = "./", gDevice = "dev.new", ...)
}
\arguments{
\item{object}{a cellData object}

\item{genemap}{the mapping file of all samples}

\item{email}{the email for the DAVID WebService registration}

\item{type}{type=c("all","high","low"), all hits or only the high/low
expressed hits to analysis}

\item{david.terms}{a list of DAVID annotation categories

                   Default: list(GO=c("GOTERM_BP_FAT","GOTERM_MF_FAT",
"GOTERM_CC_FAT"),PATHWAY=c( "KEGG_PATHWAY","REACTOME_PATHWAY","BIOCARTA"))}

\item{count}{numeric, number of items}

\item{threshold}{numeric, threshold in the DAVID analysis}

\item{showTermNum}{numeric, number of terms to show in the figure}

\item{showTermBy}{character, the order of the DAVID chart,
showTermBy is one of the column names of the chart.}

\item{showTermCol}{character, color of the bars in the figure}

\item{file}{the path of the DAVID report to generate to.
file=NA disables report generation.}

\item{outpath}{a character string naming the location the figures to
generate to}

\item{gDevice}{the graphic device (default: dev.new).
gDevice=NA generates no figure}

\item{...}{arguments of the graphic device}
}
\value{
a data frame of the DAVID report with annotated well IDs
and gene symbols
}
\description{
Performs function analysis using DAVID Web Service
}
\examples{
data(demoCell)
data(genemap)
## Please register your email at DAVID WebService
chart <- cellSigAnalysis(oneCell,genemap,
                        email="cliu@sjtu.edu.cn",type="all",
                        file=NA)
colnames(chart)

}

