% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-methods.R
\docType{methods}
\name{cellSigPlot}
\alias{cellSigPlot}
\title{Hits volcano plot}
\usage{
cellSigPlot(object, outpath = getOption("opm.outpath"),
  color.highlight = getOption("opm.sig.color.highlight"),
  color.background = getOption("opm.sig.color.background"),
  highlight.label = NULL,
  highlight.label.color = getOption("opm.sig.label.color"), ...)
}
\arguments{
\item{object}{a cellData object}

\item{outpath}{diretory of the output figures}

\item{color.highlight}{a character specifying the color of the hits}

\item{color.background}{a character specifying the color of the other samples}

\item{highlight.label}{a vector of characters specifying the names of the
samples to be highlighted, with the names are the "barcode:wellID".}

\item{highlight.label.color}{a character specifying the color of the labels}

\item{...}{arguments of the graphic device and ggplot2}
}
\value{
Invisibly an object of \code{ggplot}
}
\description{
Visualizes hits by volcano plot.
}
\details{
Users can highlight a certain samples during plotting.
}
\examples{
data(demoCell)
op <- options("device")
options("device" = "png")
labels <- c("Axin1")
names(labels) <- c("DSIMGA04:C07")
cellSigPlot(oneCell, highlight.label = labels, outpath = tempdir())
options(op)
}

