% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData-methods.R
\docType{methods}
\name{cellViz}
\alias{cellViz}
\title{Data visualization}
\usage{
cellViz(object, data.type = c("raw", "norm"), plot = c("heatmap",
  "boxplot"), outpath = getOption("opm.outpath"), multiplot = FALSE,
  plateID = NULL, tag = NULL, ctr.excluded = TRUE, ...)
}
\arguments{
\item{object}{a cellData object}

\item{data.type}{c("raw", "norm), visualizing both types by default}

\item{plot}{c("heatmap","boxplot")}

\item{outpath}{directory of output figures, default: getOption("opm.outpath")}

\item{multiplot}{logical, the output images are placed in one figure or not}

\item{plateID}{numeric or character}

\item{tag}{character, unique tag for one figure}

\item{ctr.excluded}{logical, if controls are included in the visualization}

\item{...}{other arguments for graphical devices and \code{pheatmap}}
}
\value{
Invisibly a list of the values returned by \code{pheatmap} and
ggplot2 function for boxplot
}
\description{
Visualize data by heatmap or boxplot.
}
\details{
By visualizing the raw data, users can observe the batch effects as
a large region of distinguishing color in heatmap or
biased distribution by boxplots. Users can also visualize thr normalized data
for comparison.
}
\examples{

data(demoCell)
op <- options("device")
options("device" = "png")
cellViz(oneCell, data.type = c("raw", "norm"), plateID = 1:6, outpath = tempdir())
cellViz(oneCell, data.type = c("raw", "norm"), plateID = 1, outpath = tempdir())
options(op)

}

