% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{demoData}
\alias{demoData}
\alias{genemap}
\alias{oneCell}
\alias{platemap}
\title{Examples of tables and cellData objects}
\value{
platemap: a data frame

genemap:a data frame

oneCell: a \code{cellData} object
}
\description{
Examples of tables and cellData objects
}
\section{platemap}{

\describe{
\item{Description}{The experiment information of each Columbus
analysis report. This table is required only if the report formats are not
standarded. See \code{\link{loadAll}} for more information.}
\item{Format}{data.frame with the following required column names:
\describe{
\item{\code{file.name}:}{character, the name of the report.}
\item{\code{format.type}:}{character, only ''Tab'' and ''Matrix'' are supported in
the current version.}
\item{\code{Barcode}:}{character, the barcode of the plates.}
\item{\code{rep.id}:}{character, the ID to distinguish the replicated plates.}
\item{\code{path}:}{character, the full path of the report.}
}
}
}
}

\section{genemap}{

\describe{
\item{Description}{The gene names and informations of each well.
The correponding file is generated during the design of assays.
Required for function \code{\link{GenerateReport}}
and \code{\link{cellSigAnalysis}}.}
\item{Format}{data.frame with the following required column names:
\describe{
\item{\code{Barcode}:}{character, the barcode of the plates.}
\item{\code{Well}:}{character, the well ID.}
\item{\code{GeneSymbol}:}{character, the annotated gene names of the well.}
\item{\code{Gene.ID}:}{character, Entrez gene ID.}
}}
}
}

\section{oneCell}{

\describe{
\item{Description}{
\code{oneCell} is a \code{cellData} object used in the examples of the
package.}
}
}
\examples{
data(platemap)
str(platemap)
data(genemap)
str(genemap)
data(demoCell)
oneCell
}
\keyword{data}

