% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expData-class.R
\docType{class}
\name{expData-class}
\alias{[,expData,ANY,ANY,ANY-method}
\alias{[,expData-method}
\alias{dataLoad}
\alias{dataLoad,expData-method}
\alias{expData}
\alias{expData,expData-method}
\alias{expData-class}
\alias{show,expData-method}
\title{The expData class}
\description{
The expData class is a container to store data imported from one
Columbus system report

Constructor method of expData class.

Show method
}
\section{Slots}{

\describe{
\item{\code{name}}{character, the plate ID (barcode-replicateID), e.g. DSIMGA03-s1.}

\item{\code{path}}{character, the path of the Columbus system report.}

\item{\code{rep.id}}{character, replicateID, e.g. s1.}

\item{\code{exp.id}}{character, barcode, e.g. DSIMGA03.}

\item{\code{data}}{a list of vectors,
the vectorized raw data matrix of one plate of each type.}

\item{\code{format}}{character, format of the Columbus system report.}

\item{\code{wellID}}{a character vector, the well IDs.}
}}
\section{Methods}{

\describe{
\item{Constructor}{
\code{expData(name, path, rep.id, exp.id, format)}.
}
\item{Show}{signature(object="expData"). Displays object content as text.}
\item{Accessor}{x[i]. \code{x}: an expData object;
\code{i}: character, a expData slot name.}
\item{dataLoad}{\code{dataLoad(object,data,wellID)}}
}
}
\examples{
onePlate <- expData(name = "130504-s1-02.txt",
         path = file.path( system.file("Test",package = "OperaMate"),
                           "Matrix", "130504-s1-02.txt" ),
       rep.id = "s1",
       exp.id = "DSIMGA02",
       format = "Matrix")
onePlate
onePlate["name"]
}

