% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateReport}
\alias{generateReport}
\title{Report generation}
\usage{
generateReport(lstCells, genemap = NULL, verbose = FALSE, file = NULL,
  outpath = getOption("opm.outpath"), plot = TRUE, ...)
}
\arguments{
\item{lstCells}{a list of cellData objects}

\item{genemap}{a data frame, the well-gene specification table}

\item{verbose}{logical, detailed data will be provided if TRUE}

\item{file}{the path of the file to generate to}

\item{outpath}{a character string naming the location the figures to
generate to}

\item{plot}{if TRUE, plot barplot}

\item{...}{arguments of the graphic device}
}
\value{
a data frame with annotated information of each well
}
\description{
Summarizes all results in the list of \code{cellData} objects,
and writes out a report to file.
}
\details{
This function summarizes the information from all \code{cellData} objects,
and visualizes the number of the hists if required.
}
\examples{
data(demoCell)
genemap <- read.csv(file.path(system.file("Test", package = "OperaMate"),
"demoData", "genemap.csv"), stringsAsFactors = FALSE)
report <- generateReport(list(oneCell), genemap, verbose = FALSE,
plot = FALSE)
str(report)

}

