% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expData-methods.R
\docType{methods}
\name{loadAll}
\alias{loadAll}
\title{Data importing}
\usage{
loadAll(cellformat, datapath = "./", platemap = NULL, prefix = "DSIMGA")
}
\arguments{
\item{cellformat}{character specifying the format of the reports}

\item{datapath}{character specifying the location of the reports}

\item{platemap}{data frame, required when cellformat = "Others".
See an example as \code{\link{platemap}}.}

\item{prefix}{character, the Barcode prefix, e.g. "DSIMGA"}
}
\value{
a list of \code{expData} objects
}
\description{
Initializes a list of \code{expData} objects from the Columbus system
reports.
}
\details{
"Tab" and "Matrix" are corresponding to table and matrix format in the
Columbus system report. They requires all reports of the same format, in
the same location, and with the file names as  *-replicateID-plateID.txt.
Otherwise,choose "Others" and specify the parameter of each file in
\code{platemap}. If the reports are of other cellformats, you can specify
its cellformat and rewrite the function \code{parseTemplete} to import the
data seperately. See the example to find the format for the slot \code{data}
and \code{wellID}.
}
\examples{
data(platemap)
platemap$path <- file.path(system.file("Test",package = "OperaMate"),
platemap$path)
## Not consistent format
lstPlates <- loadAll(cellformat="Others",platemap=platemap)
## Consistent format
## lstPlates <- loadAll(cellformat="Tab",datapath=dirname(platemap$path[1]))
## User defined format
## lstPlates <- loadAll(dirname(platemap$path[1]),cellformat="usrDef")
## lstPlates <- lapply(lstPlates,function(onePlate) parseTemplete(onePlate))
lstPlates[[1]]
}

