% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{operaMate}
\alias{operaMate}
\title{Data process and analysis pipeline}
\usage{
operaMate(configFile, gDevice = "png", ...)
}
\arguments{
\item{configFile}{the location of the file specifying all parameters}

\item{gDevice}{the graphics device}

\item{...}{addition arguments for graphics devices}
}
\value{
a list of three components:  a list of cellData objects, the
annotated table of each well, and the enrichment analysis table
}
\description{
A systematical pipeline for opera data importing, normalization, quality
control, hit detection, analysis, and visualization.
}
\examples{
configFile <- file.path(system.file("Test", package = "OperaMate"),
"demoData", "demoParam.txt")
operaReport <- operaMate(configFile, gDevice = "png")
head(operaReport$report)

}

