% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData-methods.R
\docType{methods}
\name{parseTemplete}
\alias{parseTemplete}
\title{Data extraction from one report}
\usage{
parseTemplete(onePlate)
}
\arguments{
\item{onePlate}{an expData object}
}
\value{
an \code{expData} object with initialized slot \code{data}.
}
\description{
Extracts data in the report to the slot \code{data} in the \code{expData}
object. An inner function of \code{\link{loadAll}}.
}
\examples{
data(platemap)
platemap$path <- file.path(system.file("Test",package = "OperaMate"),
platemap$path)
lstPlates <- loadAll(cellformat="Tab",datapath=dirname(platemap$path[1]))
onePlate <- lstPlates[[1]]
parseTemplete(onePlate)
}

